Submission version
after having the class inheriting from mono and polymorphism in using the changecolor() method (also in using start() and update(), if you want), we've added a new rotating method in update, different in each class (inheritance and polymorphism, again), we have created a separated method to call in order to make the object rotating (abstraction). finally, we put getters/setters to the color variable in the shape script, for encapsulation. (we also added a rotating camera and an audio clip).